Backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pandas>=2.0.0
pydantic>=2.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
openai>=1.0.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
PyJWT>=2.8.0
